Start refactoring parser-related functions
