+++ main.c
@@ -8,7 +8,7 @@
 
 #define MAXLINE 256
 
-enum { A_COMMAND, C_COMMAND, L_COMMAND, JUMP, COMP, DEST };
+//enum { A_COMMAND, C_COMMAND, L_COMMAND, JUMP, COMP, DEST };
 
 const char *dest_keys[] = { "M", "D", "MD", "A", "AM", "AD", "AMD" };
 const int   dest_vals[] = { 1, 2, 3, 4, 5, 6, 7 };
@@ -27,22 +27,6 @@ const char  comp_vals[] = {
 };
 
 // Conrad: consider putting all above constants in their own .c file
-
-// Function:    get_command_type
-// Description: This function takes in a cleaned line from an .asm file and returns
-//              the instruction type: A-instruction, C-instruction, or L-instruction
-// Parameters:
-//              - line: cleaned .asm line
-// Returns:     Command-type enum
-int get_command_type(char *line)
-{
-    if (line[0] == '@')
-        return A_COMMAND;
-    else if (line[0] == '(')
-        return L_COMMAND;
-    else
-        return C_COMMAND;
-}
 
 // Function:    itob
 // Description: Encodes an integer to a bitstring of length `len`
@@ -83,37 +67,6 @@ void build_A_COMMAND(char *line_in, char *line_out, LinkedList *symbols, int *de
         if (i == *default_val)
             (*default_val)++;
         itob(i, line_out, 16);
-    }
-}
-
-// Function:    tokenize
-// Description: Tokenizes cleaned C-instruction line into its three component parts
-// Parameters:
-//              line:   pointer to line to tokenize
-//              comp:   pointer to computation token (mutated by function)
-//              dest:   pointer to destination token (mutated by function)
-//              jump:   pointer to jump token (mutated by function)
-//  Returns:    void
-void tokenize(char *line, char *comp, char *dest, char *jump) 
-{
-    char *equal_sign, *semicolon;
-
-    if ((equal_sign = strchr(line, '='))) {
-        *equal_sign = '\0';
-        strcpy(dest, line);
-        strcpy(comp, equal_sign + 1);
-        if ((semicolon = strchr(comp, ';'))) {
-            *semicolon = '\0';
-            strcpy(jump, semicolon + 1);
-        } else {
-            strcpy(jump, "");
-        }
-    } else {
-        semicolon = strchr(line, ';');
-        *semicolon = '\0';
-        strcpy(dest, "");
-        strcpy(comp, line);
-        strcpy(jump, semicolon + 1);
     }
 }
